# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Workflow Agent Service API
  version: 0.1.0
  description: Versioned public HTTP contract for the workflow and Agent platform.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /
security:
  - cookieAuth: []
...
paths:
...
  /workflow/process/start:
    post:
      tags: [Workflow]
      operationId: startProcess
      summary: Start a process instance
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartProcessRequest'
      responses:
        '200':
          description: Process started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartProcessApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /workflow/process/interaction:
    post:
      tags: [Workflow]
      operationId: getStartProcessInteraction
      summary: Generate the tenant-facing data contract for the next Agent frontier
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProcessInteractionRequest' }
      responses:
        '200': { description: Generated interaction contract, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessInteractionApiResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /workflow/task/approve:
    post:
      tags: [Workflow]
      operationId: approveTask
      summary: Approve and complete a user task
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskActionRequest'
      responses:
        '200':
          description: Task completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskActionApiResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /workflow/process/participant-requirements:
    post:
      tags: [Workflow]
      operationId: getStartParticipantRequirements
      summary: Resolve participants required to start a process
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProcessParticipantRequirementsRequest' }
      responses:
        '200': { description: Participant requirements, content: { application/json: { schema: { $ref: '#/components/schemas/ParticipantRequirementListApiResponse' } } } }
...
  /workflow/process/status:
    get:
      tags: [Workflow]
      operationId: getProcessStatus
      summary: Read process status
      parameters: [{ $ref: '#/components/parameters/ProcessInstanceId' }]
      responses:
        '200': { description: Process status, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessStatusApiResponse' } } } }
  /workflow/task/status:
    get:
      tags: [Workflow]
      operationId: getTaskStatus
      summary: Read task status
      parameters: [{ $ref: '#/components/parameters/TaskId' }]
      responses:
        '200': { description: Task status, content: { application/json: { schema: { $ref: '#/components/schemas/TaskApiResponse' } } } }
  /workflow/task/participant-requirements:
    post:
      tags: [Workflow]
      operationId: getTaskParticipantRequirements
      summary: Resolve participants required for a task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskParticipantRequirementsRequest' }
      responses:
        '200': { description: Participant requirements, content: { application/json: { schema: { $ref: '#/components/schemas/ParticipantRequirementListApiResponse' } } } }
  /workflow/task/interaction:
    post:
      tags: [Workflow]
      operationId: getTaskInteraction
      summary: Generate the tenant-facing data contract for an Agent reached after task approval
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskInteractionRequest' }
      responses:
        '200': { description: Generated interaction contract, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessInteractionApiResponse' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /workflow/task/reject:
    post:
      tags: [Workflow]
      operationId: rejectTask
      summary: Reject a user task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskActionRequest' }
      responses:
        '200': { description: Task rejected, content: { application/json: { schema: { $ref: '#/components/schemas/TaskActionApiResponse' } } } }
  /workflow/task/transfer:
    post:
      tags: [Workflow]
      operationId: transferTask
      summary: Transfer a user task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransferTaskRequest' }
      responses:
        '200': { description: Task transferred, content: { application/json: { schema: { $ref: '#/components/schemas/TaskApiResponse' } } } }
  /workflow/process/definition/diagram:
    get:
      tags: [Workflow]
      operationId: getProcessDefinitionDiagram
      summary: Read process definition diagram data
      parameters:
        - { $ref: '#/components/parameters/ProcessDefinitionKey' }
        - { name: version, in: query, schema: { type: integer } }
        - { name: processDefinitionId, in: query, schema: { type: string } }
      responses:
        '200': { description: Definition diagram data, content: { application/json: { schema: { $ref: '#/components/schemas/ProcessDefinitionDiagramApiResponse' } } } }
...
components:
  parameters:
...
    ProcessDefinitionKey:
      name: processDefinitionKey
      in: query
      required: true
      schema: { type: string }
    ProcessInstanceId:
      name: processInstanceId
      in: query
      required: true
      schema: { type: string }
    TaskId:
      name: taskId
      in: query
      required: true
      schema: { type: string }
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: WORKFLOW_ACCESS_TOKEN
  schemas:
...
    ApiResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string }
        message: { type: string }
        data: {}
        traceId: { type: string }
...
    StartProcessApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/StartProcessResult' } }
    TaskActionApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/TaskActionResult' } }
    TaskApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/TaskView' } }
    ParticipantRequirementListApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { type: array, items: { $ref: '#/components/schemas/ParticipantRequirement' } } }
    ProcessInteractionApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ProcessInteraction' } }
    ProcessStatusApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ProcessStatus' } }
...
    ProcessDefinitionDiagramApiResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties: { data: { $ref: '#/components/schemas/ProcessDefinitionDiagram' } }
...
    TaskView:
      type: object
      required: [taskId, processInstanceId, taskDefinitionKey, status, candidateUsers, candidateGroups]
      properties:
        taskId: { type: string }
        processInstanceId: { type: string }
        processDefinitionId: { type: string }
        taskDefinitionKey: { type: string }
        taskName: { type: string, nullable: true }
        assignee: { type: string, nullable: true }
        candidateUsers: { type: array, items: { type: string } }
        candidateGroups: { type: array, items: { type: string } }
        status: { type: string }
        startTime: { type: string, format: date-time, nullable: true }
        endTime: { type: string, format: date-time, nullable: true }
        deleteReason: { type: string, nullable: true }
    StartProcessResult:
      type: object
      required: [processInstanceId, processDefinitionId, processDefinitionKey, activeTasks]
      properties:
        processInstanceId: { type: string }
        processDefinitionId: { type: string }
        processDefinitionKey: { type: string }
        businessKey: { type: string, nullable: true }
        activeTasks: { type: array, items: { $ref: '#/components/schemas/TaskView' } }
    TaskActionResult:
      type: object
      required: [completedTaskId, processInstanceId, processEnded, nextTasks]
      properties:
        completedTaskId: { type: string }
        processInstanceId: { type: string }
        processEnded: { type: boolean }
        nextTasks: { type: array, items: { $ref: '#/components/schemas/TaskView' } }
    ParticipantRequirement:
      type: object
      required: [activityId, activityName, assignmentType, multiple, required]
      properties:
        activityId: { type: string }
        activityName: { type: string }
        assignmentType: { type: string }
        multiple: { type: boolean }
        required: { type: boolean }
    ProcessInteraction:
      type: object
      required: [fields, agentActivityIds]
      properties:
        fields: { type: array, items: { $ref: '#/components/schemas/InteractionDataField' } }
        agentActivityIds: { type: array, items: { type: string } }
    InteractionDataField:
      type: object
      required: [variablePath, label, dataType, required, agentActivityId, agentActivityName, agentInputPath]
      properties:
        variablePath: { type: string }
        label: { type: string }
        description: { type: string, nullable: true }
        dataType: { type: string, enum: [string, number, integer, boolean, object, array] }
        format: { type: string, nullable: true }
        required: { type: boolean }
        currentValue: {}
        agentActivityId: { type: string }
        agentActivityName: { type: string }
        agentInputPath: { type: string }
    ProcessStatus:
      type: object
      required: [processInstanceId, processDefinitionId, status, activeTasks]
      properties:
        processInstanceId: { type: string }
        processDefinitionId: { type: string }
        businessKey: { type: string, nullable: true }
        status: { type: string }
        activeTasks: { type: array, items: { $ref: '#/components/schemas/TaskView' } }
...
    ProcessDefinitionDiagram:
      type: object
      required: [processDefinitionId, processDefinitionKey, version, deploymentId, active, bpmnXml, tenantId]
      properties:
        processDefinitionId: { type: string }
        processDefinitionKey: { type: string }
        processDefinitionName: { type: string, nullable: true }
        version: { type: integer }
        deploymentId: { type: string }
        deployTime: { type: string, format: date-time, nullable: true }
        active: { type: boolean }
        bpmnXml: { type: string }
        tenantId: { type: string }
...
    StartProcessRequest:
      type: object
      required: [processDefinitionKey]
      properties:
        processDefinitionKey: { type: string }
        processDefinitionId: { type: string }
        businessKey: { type: string }
        variables: { type: object, additionalProperties: true }
...
    ProcessParticipantRequirementsRequest:
      type: object
      required: [processDefinitionKey]
      properties:
        processDefinitionKey: { type: string }
        processDefinitionId: { type: string }
        variables: { type: object, additionalProperties: true }
    ProcessInteractionRequest:
      type: object
      required: [processDefinitionKey]
      properties:
        processDefinitionKey: { type: string }
        processDefinitionId: { type: string }
        variables: { type: object, additionalProperties: true }
    TaskInteractionRequest:
      type: object
      required: [taskId]
      properties:
        taskId: { type: string }
        variables: { type: object, additionalProperties: true }
    TaskParticipantRequirementsRequest:
      type: object
      required: [taskId]
      properties:
        taskId: { type: string }
        variables: { type: object, additionalProperties: true }
    TransferTaskRequest:
      type: object
      required: [taskId, targetAssignee]
      properties:
        taskId: { type: string }
        targetAssignee: { type: string }
        comment: { type: string }
...
    TaskActionRequest:
      type: object
      required: [taskId]
      properties:
        taskId: { type: string }
        comment: { type: string }
        variables: { type: object, additionalProperties: true }
...
  responses:
...
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
...